Refactor import paths for admin components to improve consistency
